requirements.txt
Flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2 